requirements.txt
discord.py
imapclient
beautifulsoup4
requests
python-dotenv